Enabling assertions with GNU Classpath 0.99.
The first use of class VMConstructor is inside the JVM not in jited code. So, the is not ready. Solution??? The class is marked as ready by hand. The same for VMField and VMMethod
(cherry picked from commit 24cbf4f6b2d9bfed272ae82a9d6480ba27b66d5f)

git-svn-id: http://llvm.org/svn/llvm-project/vmkit/trunk@180311 91177308-0d34-0410-b5e6-96231b3b80d8